requirements.txt
python-dotenv
groq
pyaudio
speechrecognition
pydub
gtts
elevenlabs
gradio